requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-multipart==0.0.6
insightface==0.7.3
onnxruntime==1.17.1
opencv-python==4.9.0.80
numpy==1.26.3
pydantic-settings==2.1.0
websockets==12.0
